Almost done! There is exception from concurrent process, constants in a code and not very good design

+++ src/jswing/RectangularSilhouette.java
@@ -7,8 +7,8 @@ public class RectangularSilhouette extends Silhouette implements pontoCheckable{
     }
 
     public boolean checkPonto(Ponto p){
-        float x = p.getX();
-        float y = p.getY();
+        double x = p.getX();
+        double y = p.getY();
         double r = getR();
 
         return ((x > -r) && (x < 0)) && ((y > -r) && (y < 0));
